requirements.txt
fastapi==0.95.1
uvicorn==0.21.1
kafka-python==2.0.2
mysql-connector-python==8.0.33
requests==2.28.2
schedule==1.2.0
python-dotenv==1.0.0
